+++ CHANGES.rst
@@ -16,7 +16,9 @@
   (`Issue #27 <https://github.com/graingert/redshift_sqlalchemy/pull/27>`_)
 - Fix unicode issue with SORTKEY on python 2.
   (`Issue #34 <https://github.com/graingert/redshift_sqlalchemy/pull/34>`_)
-
+- Add support for Redshift ``DELETE`` statements that refer other tables in the ``WHERE``
+  clause. (`Issue #35 <https://github.com/graingert/redshift_sqlalchemy/issues/35>`_).
+  Thanks `haleemur <https://github.com/haleemur> _`.
 
 0.1.2 (2015-08-11)
 ------------------